refactor(async-await): improvements on MockWebAuthenticationService.swift
